requirements.txt
tensorflow
fastapi
uvicorn
numpy
scikit-learn
Pillow
python-multipart
